mcu_opcodes: stronger typing for operand sizes
This commit gives the enum a name `MCU_Operand_Size` and replaces all
uses with the new type. Despite only having 2 states this value was
previously passed around as uint32_t and used implicit boolean tests to
distinguish between word and byte. This commit replaces the implicit
tests with explicit switch statements. We do not use if statements so
that each call-site is forced to explicitly handle both cases to make
usage obvious.