decoder2: initial commit
This commit consists of 150+ squashed commits implementing a faster
H8/532 VM. The new VM supports instruction caching to reduce the number
of rom loads and time spent decoding instructions.

At the time of this commit, the new VM passes the full test suite of
mk1, mk2, and jv880 romsets. It has also undergone several hundred hours
of external testing with music from various doom wads in an attempt to
find yet unimplemented instructions.

This feature is disabled by default. To enable it, cmake must be
configured with `-DNUKED_ENABLE_DECODER2=ON`. After cleanup and the
addition of tests for other romsets it will become the default
implementation in 0.7.0.