mcu_timer: fix invalid optimization
ed81222b3272e3416e740c28df595b1fee5e8935 introduced some lookup table
based optimizations that cause output to differ from upstream. This
seems to only affect JV880 output, but not in an audible way.
Regardless, this needs to be fixed so that we can add correct
integration tests for JV880.

The performance impact for other romsets is negligible.